Move the file components to a separate file entrypoint
