+++ src/file/index.ts
@@ -1,0 +1,5 @@
+export { default as FileInput, FileType } from "src/file/FileInput";
+export { default as FileInputList } from "src/file/FileInputList";
+
+export type { FileInputProps } from "src/file/FileInput";
+export type { FileInputListProps } from "src/file/FileInputList";
