+++ src/root/components/index.ts
@@ -1,15 +1,11 @@
 export { default as Artwork } from "src/root/components/Artwork";
 export { default as CollapsibleItem } from "src/root/components/CollapsibleItem";
 export { default as ExternalLink } from "src/root/components/ExternalLink";
-export { default as FileInput, FileType } from "src/root/components/FileInput";
-export { default as FileInputList } from "src/root/components/FileInputList";
 export { default as Page } from "src/root/components/Page";
 export { default as SkeletonRow } from "src/root/components/SkeletonRow";
 export { default as SwitchWithIcons } from "src/root/components/SwitchWithIcons";
 
 export type { CollapsibleItemProps } from "src/root/components/CollapsibleItem";
 export type { ExternalLinkProps } from "src/root/components/ExternalLink";
-export type { FileInputProps } from "src/root/components/FileInput";
-export type { FileInputListProps } from "src/root/components/FileInputList";
 export type { SkeletonRowProps } from "src/root/components/SkeletonRow";
 export type { SwitchWithIconsProps } from "src/root/components/SwitchWithIcons";
